// frontend/src/lib/api/DiscoveryPhoto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DiscoveryPhoto = {
  id: string;
  short_id: string;
  target: string | null;
  original_name: string | null;
  width: number | null;
  height: number | null;
  blurhash: string | null;
  appreciations_count: number;
  published_at: string | null;
  author_id: string;
  author_handle: string;
  author_display_name: string;
};